move: 동아리 관련 PreviewParameter 패키지 이동
