+++ feature/main/src/main/java/com/ku_stacks/ku_ring/main/club/compose/inner_screen/ClubListScreen.kt
@@ -50,7 +50,7 @@ import com.ku_stacks.ku_ring.main.club.compose.components.ClubTopBar
 import com.ku_stacks.ku_ring.ui.club.ClubItemColumn
 import com.ku_stacks.ku_ring.ui.club.ClubListSortButtonRow
 import com.ku_stacks.ku_ring.ui.club.ClubSortOption
-import com.ku_stacks.ku_ring.ui.club.ClubSummaryPreviewParameterProvider
+import com.ku_stacks.ku_ring.ui.club.preview.ClubSummaryPreviewParameterProvider
 import com.ku_stacks.ku_ring.ui.dialog.LoginAlertDialog
 import com.ku_stacks.ku_ring.util.showToast
